# 5x speed-up for trace.id / trace.get_id()