--- a/CHANGELOG.txt
+++ b/CHANGELOG.txt
@@ -5,7 +5,7 @@ Changes:
  - obspy.core:
    * response: allow recalculating overall sensitivity even if unit type is not
      one of VEL/ACC/DISP (see #3235)
-   * trace: 4x quicker retrieval of seed-id with trace.id / trace.get_id()
+   * trace: 5x quicker retrieval of seed-id with trace.id / trace.get_id()
  - obspy.io.nlloc:
    * set origin evaluation status to "rejected" if nonlinloc reports the
      location run as "ABORTED", "IGNORED" or "REJECTED" (see #3230)