@filefolder 2023-01-10
fwiw here they all are on my py3.11 as well

```
%timeit out = "%(network)s.%(station)s.%(location)s.%(channel)s"; out % (tr.stats)
    1.32 µs ± 4.75 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit "{}.{}.{}.{}".format( tr.stats.network, tr.stats.station, tr.stats.location, tr.stats.channel)
    342 ns ± 6.93 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit (tr.stats.network + '.' + tr.stats.station + '.' + tr.stats.location + '.' + tr.stats.channel)
    267 ns ± 2.87 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit '.'.join((tr.stats.network, tr.stats.station, tr.stats.location, tr.stats.channel))
    232 ns ± 1.05 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each) 

```