@flixha 2023-01-10
> Can you try timing:
> 
> ```python
> %timeit '.'.join((tr.stats.network, tr.stats.station, tr.stats.location, tr.stats.channel))
> ```
> 
> For me it is about 20% faster than the current implementation (# Alt 2) and is quite a bit shorter/clearer.

222 ns for me, so indeed the fastest with 20 % faster than Alt 2 - very good!. Updated times above.