@heavelock 2023-01-11
Just for sake of completeness I run those benchmarks plus an f-string solution.

```python 
%timeit out = "%(network)s.%(station)s.%(location)s.%(channel)s"; out % (tr.stats)
1.72 µs ± 36.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit "{}.{}.{}.{}".format( tr.stats.network, tr.stats.station, tr.stats.location, tr.stats.channel)
512 ns ± 1.49 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit (tr.stats.network + '.' + tr.stats.station + '.' + tr.stats.location + '.' + tr.stats.channel)
379 ns ± 2.07 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit '.'.join((tr.stats.network, tr.stats.station, tr.stats.location, tr.stats.channel))
296 ns ± 0.939 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit f"{tr.stats.network }.{tr.stats.station}.{tr.stats.location}.{tr.stats.channel}"
278 ns ± 0.958 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

It seems that the f-string is even faster :D